Restore backward compatibility of get_alipf_arrays()
This function accidentally broke backward compatibility, since the
pscore array is of type 'int' now, while it has been of type 'short
int'. Compatibility is now restored by keeping a copy of the 'int'
version in vrna_fold_compound_t

+++ src/ViennaRNA/alifold.h
@@ -9,6 +9,7 @@
 #include <ViennaRNA/ribo.h>
 #include <ViennaRNA/mfe.h>
 #include <ViennaRNA/part_func.h>
+#include <ViennaRNA/aln_util.h>
 #include <ViennaRNA/boltzmann_sampling.h>
 
 #ifdef DEPRECATION_WARNINGS
@@ -423,7 +424,7 @@ DEPRECATED(int get_alipf_arrays(short ***S_p,
                      FLT_OR_DBL **qm_p,
                      FLT_OR_DBL **q1k_p,
                      FLT_OR_DBL **qln_p,
-                     int **pscore));
+                     short **pscore));
 
 
 /**
